requirements.txt
awsebcli==3.10.2
blessed==1.14.2
cement==2.8.2
certifi==2018.1.18
chardet==3.0.4
click==6.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
Flask==0.12.2
github3.py==1.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
pathspec==0.5.0
python-dateutil==2.7.0
requests==2.18.4
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.22
websocket-client==0.40.0
Werkzeug==0.14.1
